refactor: hook 디렉토리 위치 정상화
