+++ REQUIREMENTS_STEP2.md
@@ -19,4 +19,4 @@
 - [x] 카드 목록은 최신순(내림차순)으로 정렬된다.
 - [x] 목록 최상단에 +을 누르면 카드 추가 페이지로 이동한다.
 - 카드를 클릭하면, 카드 별칭 수정(카드 추가 완료 페이지)로 이동한다.
-- [x] 카드를 삭제할 수 있다.
+- [x]카드를 삭제할 수 있다.
